get hero and gif from api
