+++ src/Pages/ProfilePage/ProfilePage.jsx
@@ -1,5 +1,5 @@
 import Profile from '../../components/Profile/Profile'
-import Map from '../../components/Map/Map.tsx'
+import Map from '../../components/Map/Map.jsx'
 import './ProfilePage.scss'
 function ProfilePage ({giveCoords, coords, decodedToken, getLoginId}) {
