--- src/main.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
-import App from './App.jsx'
-
-
-
-ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
-    <App />
-)
